reorganize PR
