+++ avaframe/ana5Hybrid/hybridModel_com1DFACfg.ini
@@ -1,5 +1,5 @@
 ### Config File - This file contains the changed settings for the simulation run
-## It replaces the local_com1DFACfg.ini 
+## It replaces the local_com1DFACfg.ini
 
 [GENERAL]
 #++++++++++++++++ Simulation type
@@ -25,7 +25,7 @@ stopCritType = massFlowing
 # if based on massFlowing, specify the velocity threshold for flowing mass (m/s)
 uFlowingThreshold = 0.0
 # stopCriterion (stops when massFlowing<0.01*peakMassFlowing or ke<0.01*pke)
-stopCrit = 0.0
+stopCrit = 0.01
 
 #+++++++++++++SPH parameters
 # SPH gradient option
@@ -88,7 +88,7 @@ meshCellSize = 5
 
 #+++++++++++++Flow model parameters+++++++++
 # subgridMixingFactor
-subgridMixingFactor = 10.
+subgridMixingFactor = 20.
 # curvature acceleration coefficient
 # take curvature term into account in the gravity acceleration term
 # 0 if deactivated, 1 if activated
@@ -102,7 +102,7 @@ explicitFriction = 1
 frictModel = Coulomb
 #+++++++++++++General Friction parameters
 # tan of bed friction angle used for: samosAT, Coulomb, Voellmy
-mu = 0.3
+mu = 0.5822150487715261
 
 [FLAGS]
 # True if dev release areas
