+++ avaframe/out3Plot/plotUtils.py
@@ -130,7 +130,7 @@ cmapP = copy.copy(cmapCrameri.hawaii.reversed())
 levD = [0.5, 1.0, 2.0, 3.0, 4.0, 5.0]
 # Lajolla color map
 colorsD = ["#FCFFC9", "#EBCE7B", "#DE9529", "#BE5A32", "#7F2B3F", "#1D0B14"]
-cmapD = copy.copy(cmapCrameri.lajolla.copy())
+cmapD = copy.copy(cmapCrameri.lajolla)
 
 # multi sequential colormap for speed
 levS = [1, 5, 10, 15, 20, 25, 30]
@@ -140,11 +140,12 @@ colorsS = ['#FFCEF4', '#FFA7A8', '#C19A1B', '#578B21', '#007054', '#004960',
 cmapS = copy.copy(cmapCrameri.batlow.reversed())
 
 # multi sequential colormap for Travel Angle
-levTA = [19, 20, 21, 22, 25, 30, 40]
+# levTA = [18, 19, 20, 21, 22, 22, 23]
+levTA = [28, 29, 30, 31, 32, 33, 34]
 # Batflow color map
 colorsTA = ['#FFCEF4', '#FFA7A8', '#C19A1B', '#578B21', '#007054', '#004960',
            '#201158']
-cmapTA = copy.copy(cmapCrameri.lapaz.reversed())
+cmapTA = copy.copy(cmapCrameri.lapaz)
 
 # colormap used if no resType provided
 cmapNN = copy.copy(cmapCrameri.imola.reversed())
@@ -382,7 +383,7 @@ def saveAndOrPlot(pathDict, outFileName, fig):
     return outPath
 
 
-def constrainPlotsToData(inputData, cellSize, extentOption=False):
+def constrainPlotsToData(inputData, cellSize, extentOption=False, constrainedData=False):
     """ constrain inut raster dataset to where there is data plus buffer zone
 
         Parameters
@@ -418,15 +419,21 @@ def constrainPlotsToData(inputData, cellSize, extentOption=False):
         colsMax = inputData.shape[1]
 
     if extentOption:
-        dataConstrained = inputData[rowsMin:rowsMax+1, colsMin:colsMax+1]
         rowsMinPlot = rowsMin*cellSize
         rowsMaxPlot = (rowsMax+1)*cellSize
         colsMinPlot = colsMin*cellSize
         colsMaxPlot = (colsMax+1)*cellSize
-
-        return rowsMinPlot, rowsMaxPlot, colsMinPlot, colsMaxPlot, dataConstrained
+        if constrainedData:
+            dataConstrained = inputData[rowsMin:rowsMax+1, colsMin:colsMax+1]
+            return rowsMinPlot, rowsMaxPlot, colsMinPlot, colsMaxPlot, dataConstrained
+        else:
+            return rowsMinPlot, rowsMaxPlot, colsMinPlot, colsMaxPlot
     else:
-        return rowsMin, rowsMax, colsMin, colsMax
+        if constrainedData:
+            dataConstrained = inputData[rowsMin:rowsMax+1, colsMin:colsMax+1]
+            return rowsMin, rowsMax, colsMin, colsMax, dataConstrained
+        else:
+            return rowsMin, rowsMax, colsMin, colsMax
 
 
 def addColorBar(im, ax2, ticks, myUnit, title='', extend='neither', pad=0.05):
